requirements.txt
Flask
requests
python-dotenv
waitress
flask-sqlalchemy
psycopg2-binary
